+++ client/src/App.tsx
@@ -14,7 +14,7 @@ import PlantLog from './views/PlantLog/PlantLog';
 import Register from './views/Register/Register';
 import Wrapper from './components/Wrapper/Wrapper';
 import KeithTestGrid from './views/KeithTestGrid/KeithTestGrid'
-import Winner from './views/Winner/Winner';
+// import Winner from './views/Winner/Winner';
 import GameBoard from './views/GameBoard/GameBoard';
 import Header from './components/Header/Header'
 import Footer from './components/Footer/Footer'
@@ -30,14 +30,14 @@ function App() {
               <Switch>
                 <Route path="/keith-dev" exact component={KeithTestGrid} />
                 <Route path="/" exact component={Login} />
-                <Route path="/challenges" exact component={Challenges} />
-                <Route path="/challengeDetail" exact component={ChallengeDetail} />
+                <Route path="/challenged" exact component={Challenged} />
+                {/* <Route path="/challengeDetail" exact component={ChallengeDetail} /> */}
                 <Route path="/community" exact component={Community} />
                 <Route path="/home" exact component={Home} />
                 <Route path="/landing" exact component={Landing} />
                 <Route path="/plant-log" exact component={PlantLog} />
                 <Route path="/register" exact component={Register} />
-                <Route path="/winner" exact component={Winner} />
+                {/* <Route path="/winner" exact component={Winner} /> */}
                 <Route path="/game-board" exact component={GameBoard} />
               </Switch>
             </div>
