+++ client/src/components/ChallengeDisplay/ChallengeDisplay.css
@@ -1,0 +1,3 @@
+div.challenge {
+    color: white;
+}
