+++ client/src/components/ChallengeDisplay/ChallengeDisplay.tsx
@@ -11,6 +11,10 @@ import IUser from '../../interfaces/IUser';
 import INewChallenge from '../../interfaces/INewChallenge';
 import IChallenge from '../../interfaces/IChallenge';
 
+import { Link } from 'react-router-dom';
+
+import './ChallengeDisplay.css'
+
 
 
 type Props = {
@@ -26,16 +30,18 @@ const ChallengeDisplay: React.FC<Props> = (props) => {
     console.log(props.currentChallenger)
 
     return (
-        <div>
+        <div className="challenge">
             <h1>Current Challenge</h1>
             <h2>{state.currentUser.nickname} Challenge Stats</h2>
-            {props.position === 1 ? <p>Current Multiplier: {props.currentChallenge!.playerOne_currentMultiplier}</p> : <p>{props.currentChallenge!.playerTwo_currentMultiplier}</p>}
-            { props.position === 1 ? <p>Current Score: {props.currentChallenge!.playerOne_currentScore}</p> : <p> {props.currentChallenge!.playerTwo_currentScore}</p>}
+            {props.position === 1 ? <p>Current Multiplier: {props.currentChallenge!.playerOne_currentMultiplier}</p> : <p>Current Multiplier: {props.currentChallenge!.playerTwo_currentMultiplier}</p>}
+            { props.position === 1 ? <p>Current Score: {props.currentChallenge!.playerOne_currentScore}</p> : <p>Current Score: {props.currentChallenge!.playerTwo_currentScore}</p>}
 
             <h2>{props.currentChallenger!.nickname} Challenge Stats</h2>
             {props.position === 2 ? <p>Current Multiplier: {props.currentChallenge!.playerOne_currentMultiplier}</p> : <p>Current Multiplier: {props.currentChallenge!.playerTwo_currentMultiplier}</p>}
             { props.position === 2 ? <p>Current Score: {props.currentChallenge!.playerOne_currentScore} </p> : <p> Current Score: {props.currentChallenge!.playerTwo_currentScore}</p>}
 
+
+            <Link to="/"> <Button variant="contained">Back to Login</Button> </Link>
         </div>
     )
